requirements.txt
# Core Frameworks
fastapi==0.120.0
uvicorn==0.38.0

# Database
sqlalchemy==2.0.44

# Data validation
pydantic==2.12.3
pydantic_core==2.41.4

# Forms & file uploads
python-multipart==0.0.20

# Authentication & Security
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
python-jose==3.3.0

# Environment variables
python-dotenv==1.2.1

# Image processing
opencv-python==4.10.0.84
pytesseract==0.3.10
numpy==1.26.4

# AI API Client
groq==0.11.0

# Optional utility
requests==2.32.3
